+++ private/build_schema/build_query_fields/index.js
@@ -3,12 +3,12 @@ const {
   GraphQLString,
   GraphQLList,
   GraphQLEnumType,
-  GraphQLNonNull,
-  GraphQLInt
+  GraphQLInt,
+  GraphQLInputObjectType
 } = require('graphql')
 const get_table_rows = require('../../network/get_table_rows')
 const abi_to_ast = require('../abi_to_ast/index.js.js')
-const generate_table_scope = require('./table_scope')
+const generate_table_entries = require('./table_entries')
 
 const index_position_enum_type = new GraphQLEnumType({
   name: 'index_position_enum_type',
@@ -25,40 +25,6 @@ const index_position_enum_type = new GraphQLEnumType({
     tenth: { value: 'tenth' }
   }
 })
-
-const table_row_arguments = {
-  scope: {
-    type: GraphQLNonNull(GraphQLString),
-    description: 'The account to which this data belongs.'
-  },
-  key_type: {
-    type: GraphQLString,
-    description:
-      'Type of key specified by index_position (for example - uint64_t or name).'
-  },
-  index_position: {
-    type: index_position_enum_type,
-    description: 'Position of the index used.'
-  },
-  encode_type: {
-    type: GraphQLString,
-    description: 'encode type.'
-  },
-  upper_bound: {
-    type: GraphQLString,
-    description:
-      'Filters results to return the first element that is greater than provided value in set.'
-  },
-  lower_bound: {
-    type: GraphQLString,
-    description:
-      'Filters results to return the first element that is not less than provided value in set.'
-  },
-  limit: {
-    type: GraphQLInt,
-    description: 'Limit number of results returned.'
-  }
-}
 
 /**
  * This function builds query fields of a GraphQL query of the Schema from an ABI of a given EOS smart contract.
@@ -83,14 +49,59 @@ function build_query_fields(ABI) {
     (acc, item) => ({
       ...acc,
       [`${Object.values(item)[0]}`]: {
-        description: `Query data from the table \`${Object.keys(item)[0]}\`.`,
+        description: `Query data from \`${Object.keys(item)[0]}\` table.`,
         type: GraphQLList(item[Object.keys(item)[0]]),
-        args: table_row_arguments,
-        resolve: async (_, args, { rpc_url, contract }) => {
-          args.table = Object.keys(item)[0]
-          args.code = contract
-
-          const { rows } = await get_table_rows(args, rpc_url)
+        args: {
+          arg: {
+            type: new GraphQLInputObjectType({
+              name: `${item[Object.keys(item)[0]]}_arg`,
+              fields: () => ({
+                scope: {
+                  type: GraphQLString,
+                  description: `The scope within the \`${
+                    item[Object.keys(item)[0]]
+                  }\`table to query data from.`,
+                  defaultValue: ''
+                },
+                index_position: {
+                  type: index_position_enum_type,
+                  description: 'Position of the index used.',
+                  defaultValue: 'primary'
+                },
+                key_type: {
+                  type: GraphQLString,
+                  description: `The key type of \`index_position\`; primary only supports i64. All others support i64, i128, i256, float64, float128, ripemd160, sha256. Special type name indicates an account name`
+                },
+                encode_type: {
+                  type: GraphQLString,
+                  description:
+                    'The encoding type of `key_type` dec for decimal encoding of (i[64|128|256], float[64|128]); hex for hexadecimal encoding of (i256, ripemd160, sha256).'
+                },
+                upper_bound: {
+                  type: GraphQLString,
+                  description:
+                    'Filters results to return the first element that is greater than provided value in set.'
+                },
+                lower_bound: {
+                  type: GraphQLString,
+                  description:
+                    'Filters results to return the first element that is not less than provided value in set.'
+                },
+                limit: {
+                  type: GraphQLInt,
+                  description: 'The maximum number of items to return'
+                }
+              })
+            })
+          }
+        },
+        resolve: async (_, { arg }, { rpc_url, contract }) => {
+          const table_arg = {
+            code: contract,
+            table: Object.keys(item)[0],
+            ...arg
+          }
+          const { rows } = await get_table_rows(table_arg, rpc_url)
           return rows
         }
       }
@@ -120,7 +131,9 @@ function build_query_fields(ABI) {
     }
   }
 
-  fields.table_scope = generate_table_scope(ABI.tables.map(({ name }) => name))
+  fields.table_entries = generate_table_entries(
+    ABI.tables.map(({ name }) => name)
+  )
 
   return fields
 }